mys/requirements.txt
-i https://pypi.org/simple
anyascii==0.3.2; python_version >= '3.3'
asgiref==3.8.1; python_version >= '3.8'
beautifulsoup4==4.12.3; python_full_version >= '3.6.0'
boto3==1.34.135; python_version >= '3.8'
botocore==1.34.135; python_version >= '3.8'
certifi==2024.6.2; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django==5.0.6; python_version >= '3.10'
django-filter==24.2; python_version >= '3.8'
django-modelcluster==6.3; python_version >= '3.8'
django-permissionedforms==0.1; python_version >= '3.7'
django-storages==1.14.3; python_version >= '3.7'
django-taggit==5.0.1; python_version >= '3.8'
django-treebeard==4.7.1; python_version >= '3.8'
djangorestframework==3.15.2; python_version >= '3.8'
draftjs-exporter==5.0.0
et-xmlfile==1.1.0; python_version >= '3.6'
filetype==1.2.0
gunicorn==22.0.0; python_version >= '3.7'
idna==3.7; python_version >= '3.5'
jmespath==1.0.1; python_version >= '3.7'
l18n==2021.3
laces==0.1.1; python_version >= '3.8'
openpyxl==3.1.4; python_version >= '3.8'
packaging==24.1; python_version >= '3.8'
pillow==10.3.0; python_version >= '3.8'
pillow-heif==0.16.0
polib==1.2.0
psycopg2-binary==2.9.9; python_version >= '3.7'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-decouple==3.8
pytz==2024.1
requests==2.32.3; python_version >= '3.8'
s3transfer==0.10.2; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.5; python_version >= '3.8'
sqlparse==0.5.0; python_version >= '3.8'
telepath==0.3.1; python_version >= '3.8'
typing-extensions==4.12.2; python_version >= '3.8'
urllib3==2.2.2; python_version >= '3.10'
wagtail==6.1.2; python_version >= '3.8'
wagtail-localize==1.9; python_version >= '3.8'
willow[heif]==1.8.0; python_version >= '3.8'
